ComputeImageDeepLearningFeatures is moved to public pkg; time.Measure is added into common package.
